Move HTTP exceptions to own namespace and remove Routing namespace
