+++ qa-src/Http/Exceptions/PageNotFoundException.php
@@ -16,7 +16,7 @@
 	More about this license: http://www.question2answer.org/license.php
 */
 
-namespace Q2A\Http;
+namespace Q2A\Http\Exceptions;
 
 use Q2A\Exceptions\ErrorMessageException;
